+++ apps/angular-example/src/app/abby.ts
@@ -1,28 +1,30 @@
-import { Injectable } from '@angular/core';
-import { AbbyService } from 'abby';
-import { environment } from '../environments/environment';
+import { Injectable } from "@angular/core";
+import { AbbyService } from "abby";
+import { environment } from "../environments/environment";
 
 export const abby = {
   projectId: environment.ABBY_PROJECT_ID,
-  currentEnvironment: 'test',
+  currentEnvironment: "test",
   tests: {
     AngularTest: {
-      variants: ['A', 'B', 'C', 'D'],
+      variants: ["A", "B", "C", "D"],
     },
     NotExistingTest: {
-      variants: ['A', 'B'],
+      variants: ["A", "B"],
     },
   },
-  flags: ['AngularFlag', 'AngularFlag2', 'NotExistingFlag'],
-  apiUrl: 'http://localhost:3000/',
+  flags: { AngularFlag: "Boolean", AngularFlag2: "Boolean", NotExistingFlag: "Boolean" },
+  apiUrl: "http://localhost:3000/",
   debug: true,
-};
+} as const;
 
 @Injectable({
-  providedIn: 'root',
+  providedIn: "root",
   useExisting: AbbyService,
 })
 export class Abby extends AbbyService<
-  (typeof abby)['flags'][number],
-  keyof (typeof abby)['tests']
+  keyof (typeof abby)["flags"],
+  keyof (typeof abby)["tests"],
+  (typeof abby)["tests"],
+  (typeof abby)["flags"]
 > {}