+++ apps/angular-example/src/app/test_components/b.component.ts
@@ -1,9 +1,8 @@
-import {Component} from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'b-component',
   template: '<div>TEST B</div>',
   styles: [''],
 })
-export class BTestComponent {
-}
+export class BTestComponent {}