+++ apps/web/src/lib/abby.tsx
@@ -14,14 +14,20 @@ export const {
   withAbbyApiHandler,
   getABResetFunction,
 } = createAbby({
-  projectId: process.env.ABBY_PROJECT_ID!,
+  projectId: process.env.NEXT_PUBLIC_ABBY_PROJECT_ID!,
   currentEnvironment: process.env.VERCEL_ENV ?? process.env.NODE_ENV,
+  apiUrl: process.env.NEXT_PUBLIC_ABBY_API_URL,
   tests: {
     SignupButton: {
       variants: ["A", "B"],
     },
   },
-  flags: ["AdvancedTestStats", "showFooter", "test"],
+  flags: {
+    AdvancedTestStats: "Boolean",
+    showFooter: "Boolean",
+    test: "Boolean",
+    abc: "JSON",
+  },
 });
 
 export const AbbyDevtools = withDevtools(abbyDevtools, {});