+++ packages/angular/src/lib/abby.module.ts
@@ -1,9 +1,12 @@
-import { ModuleWithProviders, NgModule } from "@angular/core";
+import { APP_INITIALIZER, InjectionToken, ModuleWithProviders, NgModule } from "@angular/core";
 import { AbbyService } from "./abby.service";
-import { AbbyConfig } from "@tryabby/core";
+import { Abby, AbbyConfig } from "@tryabby/core";
 import { AbbyFlag } from "./flag.directive";
 import { AbbyTest } from "./test.directive";
 import { DevtoolsComponent } from "./devtools.component";
+import { F } from "ts-toolbelt";
+
+export const ABBY_CONFIG_TOKEN = new InjectionToken<AbbyConfig>("AbbyConfig");
 
 @NgModule({
   declarations: [AbbyFlag, AbbyTest, DevtoolsComponent],
@@ -15,10 +18,30 @@ export class AbbyModule {
       ngModule: AbbyModule,
       providers: [
         {
+          provide: ABBY_CONFIG_TOKEN,
+          useValue: config, // assuming 'config' is your configuration data
+        },
+        {
           provide: AbbyService,
-          useFactory: () => {
-            return new AbbyService(config);
+          useFactory: (config: AbbyConfig) => {
+            return new AbbyService(config as F.Narrow<AbbyConfig>);
           },
+          deps: [ABBY_CONFIG_TOKEN],
+        },
+        {
+          provide: APP_INITIALIZER,
+          useFactory: (abby: AbbyService) => {
+            return (): Promise<any> => {
+              return new Promise((resolve, reject) => {
+                abby.init().subscribe({
+                  next: () => resolve("Initialization Successful."),
+                  error: (err) => reject(err),
+                });
+              });
+            };
+          },
+          deps: [AbbyService],
+          multi: true,
         },
       ],
     };