+++ packages/angular/src/lib/abby.service.spec.ts
@@ -1,15 +1,13 @@
-import { TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { AbbyService } from "./abby.service";
 import { AbbyModule } from "./abby.module";
 
-import { Injectable } from "@angular/core";
+import { Component, Injectable } from "@angular/core";
 import { TestStorageService } from "./StorageService";
-// import { TestStorageService } from "./StorageService";
-// import { HttpService } from "shared/src/http";
-// import { AbbyEventType } from "@tryabby/core";
-// import { Routes } from "@angular/router";
-// import { forkJoin } from "rxjs";
+import { zip } from "rxjs";
+import { Routes } from "@angular/router";
+import { AbbyFlag } from "./flag.directive";
 
 const mockConfig = {
   projectId: "mock-project-id",
@@ -21,17 +19,29 @@ const mockConfig = {
     test2: {
       variants: ["A", "B"],
     },
+    defaultTest: {
+      variants: ["A", "B"],
+    },
   },
-  flags: ["flag1", "flag2", "overridedFlag1", "overridedFlag2", "defaultFlag"],
+  flags: {
+    flag1: "Boolean",
+    flag2: "Boolean",
+    overridedFlag1: "Boolean",
+    overridedFlag2: "Boolean",
+    defaultFlag: "Boolean",
+  },
   settings: {
     flags: {
       devOverrides: {
         overridedFlag1: true,
         overridedFlag2: false,
       },
+      defaultValues: {
+        Boolean: false,
+      },
     },
   },
-};
+} as const;
 
 const mockedData = {
   tests: [
@@ -47,19 +57,19 @@ const mockedData = {
   flags: [
     {
       name: "flag1",
-      isEnabled: true,
+      value: true,
     },
     {
       name: "flag2",
-      isEnabled: false,
+      value: false,
     },
     {
       name: "overridedFlag1",
-      isEnabled: false,
+      value: false,
     },
     {
       name: "overridedFlag2",
-      isEnabled: true,
+      value: true,
     },
   ],
 };
@@ -68,111 +78,106 @@ const mockedData = {
   useExisting: AbbyService,
 })
 export class Abby extends AbbyService<
-  (typeof mockConfig)["flags"][number],
-  keyof (typeof mockConfig)["tests"]
+  keyof (typeof mockConfig)["flags"],
+  keyof (typeof mockConfig)["tests"],
+  (typeof mockConfig)["tests"],
+  (typeof mockConfig)["flags"]
 > {}
 
 describe("AbbyService", () => {
   let service: AbbyService;
+  let fixture: ComponentFixture<TestComponent>;
 
-  beforeAll(() => {
-    TestBed.configureTestingModule({
-      imports: [AbbyModule.forRoot(mockConfig)],
+  @Component({
+    template: `
+      <div *featureFlag="'flag1'">
+        <h1>Flag 1</h1>
+      </div>
+      <div *featureFlag="'flag2'">
+        <h1>Flag 2</h1>
+      </div>
+      <div *abbyTest="{ testName: 'test2', variant: 'A' }">
+        <h2>A</h2>
+      </div>
+      <div *abbyTest="{ testName: 'test2', variant: 'B' }">
+        <h3>B</h3>
+      </div>
+    `,
+  })
+  class TestComponent {}
+
+  beforeAll(async () => {
+    const fetchSpy = spyOn(window, "fetch");
+
+    const mockedResponse = new Response(JSON.stringify(mockedData), {
+      status: 200,
+      headers: { "Content-type": "application/json" },
     });
+
+    fetchSpy.and.returnValue(Promise.resolve(mockedResponse));
+
+    await TestBed.configureTestingModule({
+      declarations: [TestComponent],
+      imports: [AbbyModule.forRoot(mockConfig)],
+    })
+      .compileComponents()
+      .then(() => {
+        fixture = TestBed.createComponent(TestComponent);
+        fixture.detectChanges();
+      });
     service = TestBed.inject(AbbyService);
   });
 
   beforeEach(() => {});
 
   it("should be created", () => {
-    const fetchSpy = spyOn(window, "fetch");
-
-    const mockedResponse = new Response(JSON.stringify(mockedData), {
-      status: 200,
-      headers: { "Content-type": "application/json" },
-    });
-
-    fetchSpy.and.returnValue(Promise.resolve(mockedResponse));
-
     expect(service).toBeTruthy();
   });
 
   it("gets the stored feature flag value using a function properly", () => {
-    const mockedResponse = new Response(JSON.stringify(mockedData), {
-      status: 200,
-      headers: { "Content-type": "application/json" },
-    });
-
-    const getFetchSpy = spyOn(window, "fetch").and.callFake(() =>
-      Promise.resolve(mockedResponse)
-    );
-
-    service.getFeatureFlagValue("flag1").subscribe((value: boolean) => {
+    service.getFeatureFlagValue("flag1").subscribe((value) => {
       expect(value).toEqual(true);
     });
 
-    service.getFeatureFlagValue("flag2").subscribe((value: boolean) => {
+    service.getFeatureFlagValue("flag2").subscribe((value) => {
       expect(value).toEqual(false);
     });
   });
 
   it("returns the correct variant", () => {
-    const fetchSpy = spyOn(window, "fetch");
-
-    const mockedResponse = new Response(JSON.stringify(mockedData), {
-      status: 200,
-      headers: { "Content-type": "application/json" },
-    });
-
-    fetchSpy.and.returnValue(Promise.resolve(mockedResponse));
-
     service.getVariant("test2").subscribe((value: string) => {
       expect(value).toEqual("A");
     });
   });
 
-  // it("should repect the default values for feature flags", () => {
-  //   service.getFeatureFlagValue("defaultFlag").subscribe((value: boolean) => {
-  //     expect(value).toEqual(false);
-  //   });
-  // });
+  fit("should respect the default values for feature flags", () => {
+    service.getFeatureFlagValue("defaultFlag").subscribe((value) => {
+      expect(value).toEqual(false);
+    });
+  });
+
+  it("should respect the default values for variants", () => {
+    service.getVariant("defaultTest").subscribe((value: string) => {
+      expect(["A", "B"]).toContain(value);
+    });
+  });
 
   it("uses the devOverrides", () => {
-    const fetchSpy = spyOn(window, "fetch");
-
-    const mockedResponse = new Response(JSON.stringify(mockedData), {
-      status: 200,
-      headers: { "Content-type": "application/json" },
+    service.getFeatureFlagValue("overridedFlag1").subscribe((value) => {
+      expect(value).toEqual(true);
     });
 
-    fetchSpy.and.returnValue(Promise.resolve(mockedResponse));
-
-    service
-      .getFeatureFlagValue("overridedFlag1")
-      .subscribe((value: boolean) => {
-        expect(value).toEqual(true);
-      });
-
-    service
-      .getFeatureFlagValue("overridedFlag2")
-      .subscribe((value: boolean) => {
-        expect(value).toEqual(false);
-      });
+    service.getFeatureFlagValue("overridedFlag2").subscribe((value) => {
+      expect(value).toEqual(false);
+    });
   });
 
   it("should use the persistedValue", () => {
     const persistedValue = "A";
     const variants = ["A", "B", "C", "D"];
 
-    const fetchSpy = spyOn(window, "fetch");
     const getSpy = spyOn(TestStorageService, "get");
     const setSpy = spyOn(TestStorageService, "set");
-    const mockedResponse = new Response(JSON.stringify(mockedData), {
-      status: 200,
-      headers: { "Content-type": "application/json" },
-    });
-
-    fetchSpy.and.returnValue(Promise.resolve(mockedResponse));
 
     getSpy.and.returnValue(persistedValue);
 
@@ -183,76 +188,82 @@ describe("AbbyService", () => {
     });
   });
 
-  // it("should notify the server with OnAct", () => {
-  //   const mockedResponse = new Response(JSON.stringify(mockedData), {
-  //     status: 200,
-  //     headers: { "Content-type": "application/json" },
-  //   });
+  it("should get the correct router variant", () => {
+    @Component({})
+    class ATestComponent {}
 
-  //   const getFetchSpy = spyOn(window, 'fetch').and.callFake(() =>
-  //     Promise.resolve(mockedResponse)
-  //   );
+    @Component({})
+    class BTestComponent {}
 
-  //   console.log(getFetchSpy);
+    @Component({})
+    class CTestComponent {}
 
-  //   const onActSpy = spyOn(service, "onAct").and.identity
+    @Component({})
+    class DTestComponent {}
 
-  //   // fetchSpy.and.returnValue(Promise.resolve(mockedResponse)).and.callThrough;
+    @Component({})
+    class FlagComponent {}
 
-  //   service.onAct("test");
+    let routes: Routes = [];
 
-  //   // expect(fetchSpy).toHaveBeenCalled();
+    zip(service.getVariant("test"), service.getFeatureFlagValue("flag1")).subscribe(
+      ([angularTest, angularFlag]) => {
+        routes = [
+          service.getRouterVariant(angularTest, {
+            path: "test",
+            outlet: "test",
+            abbyVariants: {
+              A: {
+                title: "TEST A",
+                component: ATestComponent,
+              },
+              B: {
+                title: "TEST B",
+                component: BTestComponent,
+              },
+              C: {
+                title: "TEST C",
+                component: CTestComponent,
+              },
+              D: {
+                title: "TEST D",
+                component: DTestComponent,
+              },
+            },
+          }),
+          ...(angularFlag
+            ? [
+                {
+                  path: "flag",
+                  title: "Flag",
+                  component: FlagComponent,
+                },
+              ]
+            : []),
+        ];
 
-  //   // const type = fetchSpy.calls.argsFor(0)[0].valueOf();
-  //   // fetchSpy.calls.argsFor(0)[0].type;
+        expect(routes).not.toEqual([]);
+      }
+    );
+  });
 
-  //   //expect(spy).toHaveBeenCalledWith();
+  it("returns the correct possible variant values", () => {
+    service.getVariants("test").subscribe((value: readonly string[]) => {
+      expect(value).toEqual(["A", "B", "C", "D"]);
+    });
+  });
 
-  //   //expect(type).toBe(AbbyEventType.ACT);
-  // });
+  it("directives should work", () => {
+    const compiled = fixture.nativeElement as HTMLElement;
+    let flagElement = compiled.querySelector("h1");
+    let testAElement = compiled.querySelector("h2");
+    let testBElement = compiled.querySelector("h3");
 
-  // it("should get the correct router variant", () => {
-  //   let routes: Routes = [];
-
-  //   forkJoin({
-  //     angularTest: service.getVariant("test"),
-  //     angularFlag: service.getFeatureFlagValue("flag1"),
-  //   }).subscribe(({ angularTest, angularFlag }) => {
-  //     routes = [
-  //       service.getRouterVariant(angularTest, {
-  //         path: "test",
-  //         outlet: "test",
-  //         abbyVariants: {
-  //           A: {
-  //             title: "TEST A",
-  //             component: "ATestComponent" as any,
-  //           },
-  //           B: {
-  //             title: "TEST B",
-  //             component: "BTestComponent" as any,
-  //           },
-  //           C: {
-  //             title: "TEST C",
-  //             component: "CTestComponent" as any,
-  //           },
-  //           D: {
-  //             title: "TEST D",
-  //             component: "DTestComponent" as any,
-  //           },
-  //         },
-  //       }),
-  //       {
-  //         ...(angularFlag
-  //           ? {
-  //               path: "flag",
-  //               title: "Flag",
-  //               component: "FlagComponent" as any,
-  //               outlet: "flag",
-  //             }
-  //           : undefined),
-  //       },
-  //     ];
-  //     console.log(routes);
-  //   });
-  // });
+    if (flagElement) {
+      expect(flagElement.textContent).toEqual("Flag 1");
+    } else fail("querySelector is null");
+    if (testAElement) {
+      expect(testAElement.textContent).toEqual("A");
+    } else fail("querySelector is null");
+  });
 });