+++ packages/angular/src/lib/abby.service.ts
@@ -5,16 +5,28 @@ import {
   Abby,
   AbbyEventType,
   HttpService,
+  FlagValueString,
+  FlagValue,
 } from "@tryabby/core";
 import { FlagStorageService, TestStorageService } from "./StorageService";
-import { from, map, Observable, of, shareReplay, tap } from "rxjs";
+import {
+  from,
+  map,
+  merge,
+  Observable,
+  of,
+  shareReplay,
+  startWith,
+  Subject,
+  switchMap,
+  take,
+  tap,
+} from "rxjs";
 import { F } from "ts-toolbelt";
 import { Route } from "@angular/router";
+import { ABBY_CONFIG_TOKEN } from "./abby.module";
 
-type LocalData<
-  FlagName extends string = string,
-  TestName extends string = string
-> = {
+type LocalData<FlagName extends string = string, TestName extends string = string> = {
   tests: Record<
     TestName,
     ABConfig & {
@@ -22,7 +34,7 @@ type LocalData<
       selectedVariant?: string;
     }
   >;
-  flags: Record<FlagName, boolean>;
+  flags: Record<FlagName, FlagValue>;
 };
 
 @Injectable({ providedIn: "root" })
@@ -30,23 +42,22 @@ export class AbbyService<
   FlagName extends string = string,
   TestName extends string = string,
   Tests extends Record<TestName, ABConfig> = Record<TestName, ABConfig>,
-  ConfigType extends AbbyConfig<FlagName, Tests> = AbbyConfig<FlagName, Tests>
+  Flags extends Record<FlagName, FlagValueString> = Record<FlagName, FlagValueString>
 > {
-  private abby: Abby<FlagName, TestName, Tests>;
+  private abby: Abby<FlagName, TestName, Tests, Flags>;
 
   private selectedVariants: { [key: string]: string } = {};
 
-  private config: F.Narrow<AbbyConfig<FlagName, Tests>>;
+  private config: F.Narrow<AbbyConfig<FlagName, Tests, Flags>>;
 
   private projectData$?: Observable<LocalData<FlagName, TestName>>;
 
   private log = (...args: any[]) =>
     this.config.debug ? console.log(`ng.AbbyService`, ...args) : () => {};
+  private cookieChanged$ = new Subject<void>();
 
-  constructor(
-    @Inject(AbbyService) config: F.Narrow<AbbyConfig<FlagName, Tests>>
-  ) {
-    this.abby = new Abby<FlagName, TestName, Tests>(
+  constructor(@Inject(AbbyService) config: F.Narrow<AbbyConfig<FlagName, Tests, Flags>>) {
+    this.abby = new Abby<FlagName, TestName, Tests, Flags>(
       config,
       {
         get: (key: string) => {
@@ -56,6 +67,7 @@ export class AbbyService<
         set: (key: string, value: any) => {
           if (typeof window === "undefined") return;
           TestStorageService.set(config.projectId, key, value);
+          this.cookieChanged$.next();
         },
       },
       {
@@ -66,6 +78,7 @@ export class AbbyService<
         set: (key: string, value: any) => {
           if (typeof window === "undefined") return;
           FlagStorageService.set(config.projectId, key, value);
+          this.cookieChanged$.next();
         },
       }
     );
@@ -74,16 +87,19 @@ export class AbbyService<
   }
 
   public init(): Observable<void> {
-    return this.resolveData().pipe(map(() => void 0));
+    return this.resolveData().pipe(
+      tap(() => {}),
+      map(() => void 0)
+    );
   }
 
-  public getVariant<T extends TestName>(testName: T): Observable<string> {
-    this.log(`getVariant(${testName})`);
+  public getVariant<T extends keyof Tests>(testName: T): Observable<string> {
+    this.log(`getVariant(${testName as string})`);
 
     return this.resolveData().pipe(
       map((data) => this.abby.getTestVariant(testName)),
-      tap((variant) => (this.selectedVariants[testName] = variant)),
-      tap((variant) => this.log(`getVariant(${testName}) =>`, variant))
+      tap((variant) => (this.selectedVariants[testName as string] = variant)),
+      tap((variant) => this.log(`getVariant(${testName as string}) =>`, variant))
     );
   }
 
@@ -113,25 +129,32 @@ export class AbbyService<
     });
   }
 
-  public getFeatureFlagValue<
-    F extends NonNullable<ConfigType["flags"]>[number]
-  >(name: F): Observable<boolean> {
+  public getFeatureFlagValue<F extends FlagName>(name: F) {
+    console.log("getFeatureFlagValue", name, this.abby.getFeatureFlag(name));
     this.log(`getFeatureFlagValue(${name})`);
 
     return this.resolveData().pipe(
-      map((data) => data.flags[name]),
+      map((data) => this.abby.getFeatureFlag(name)),
       tap((value) => this.log(`getFeatureFlagValue(${name}) =>`, value))
     );
   }
 
   private resolveData(): Observable<LocalData<FlagName, TestName>> {
     this.projectData$ ??= from(this.abby.getProjectDataAsync()).pipe(
+      switchMap((data) => {
+        const initialData$ = of(data); // Create an observable with the initial data
+
+        return merge(this.cookieChanged$, initialData$).pipe(
+          switchMap(() => from(this.abby.getProjectDataAsync())),
+          startWith(data) // Ensure that the initial data is emitted first
+        );
+      }),
       shareReplay(1)
     );
     return this.projectData$;
   }
 
-  public getAbbyInstance(): Abby<FlagName, TestName, Tests> {
+  public getAbbyInstance(): Abby<FlagName, TestName, Tests, Flags> {
     return this.abby;
   }
 
@@ -146,4 +169,12 @@ export class AbbyService<
       ...baseRoute.abbyVariants[test],
     };
   }
+
+  public getVariants = <T extends keyof Tests>(name: T) => {
+    return this.resolveData().pipe(map((data) => this.abby.getVariants(name)));
+  };
+
+  public resetAB = <T extends keyof Tests>(name: T) => {
+    TestStorageService.remove(this.config.projectId, name as string);
+  };
 }