+++ packages/core/src/shared/schemas.ts
@@ -9,3 +9,24 @@ export const abbyEventSchema = z.object({
 });
 
 export type AbbyEvent = z.infer<typeof abbyEventSchema>;
+
+export const flagValue = z.union([
+  z.boolean(),
+  z.string(),
+  z.number(),
+  z.record(z.string(), z.unknown()),
+]);
+
+export type FlagValue = z.infer<typeof flagValue>;
+
+export type FlagValueString = "String" | "Boolean" | "Number" | "JSON";
+
+export type FlagValueStringToType<T extends FlagValueString> = T extends "String"
+  ? string
+  : T extends "Boolean"
+  ? boolean
+  : T extends "Number"
+  ? number
+  : T extends "JSON"
+  ? Record<string, unknown>
+  : never;