+++ packages/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @tryabby/react
 
+## 4.0.0
+
+### Major Changes
+
+- change feature flag behavior for multivariate flags
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/core@4.0.0
+
 ## 3.4.0
 
 ### Minor Changes